Split out fake_impls from fake_tensor

[ghstack-poisoned]
